+++ src/components/header/header.jsx
@@ -1,29 +1,53 @@
-import React, { Component} from 'react';
-
+import React,  { useState, Component} from 'react';
+import { Link } from 'react-router-dom';
 
 import NavigationData from '../data/navigationData'
-import NavigationButton from '../button/navigationButton'
 
-export default class Header extends Component {
-    constructor() {
-        super()
-        this.state = {
-            data : NavigationData,
-        }
-    }
-render () {
+export default function Header () {
+    const ButtonStyleActive = "solid 5px blueviolet";
+    const StyleForBtn =  { borderBottom : ButtonStyleActive }
 
-        const navigation = this.state.data.map(data  => <NavigationButton key={data.key} item={data.categories} data={data}/>)
+    let host = window.location.pathname;
+    host = host.split("/")[1]
+
+    
+    
+    const navigation = NavigationData.map(data  => 
+        <Link  to={data.href}  key={data.key}>
+            <button 
+                onMouseLeave={() => setTopPosition(linkClassName = disablePosition)} 
+                onMouseEnter={() => setTopPosition(linkClassName = data.title === 'products' ? activePosition : disablePosition)} 
+                style={ host === data.title ? StyleForBtn : null }
+                className="link-btn">
+                <p>{data.title}</p>
+            </button>
+        </Link>)
+
+    const navigationCategories = NavigationData[2].categories.map(data => 
+        <Link to={data.link} key={data.key}>
+            <p style={{color : host === data.link ? 'blueviolet' : null}}>{data.name}</p>
+        </Link>)
+
+    const disablePosition = '-500px' 
+    const activePosition = '70px';
         
-        return (
-                <header className="header">
-                    <h1>Games Market</h1>
-                    <div className="nav_bar">
+    let [linkClassName, setTopPosition] = useState(disablePosition)
+
+    return (
+            <header className="header">
+                <h1>Games Market</h1>
+                <div className="nav_bar">
                     {navigation}
+                    <div className='link' onMouseLeave={() => setTopPosition(linkClassName = disablePosition)} 
+                        onMouseEnter={() => setTopPosition(linkClassName = activePosition)}
+                        style={{
+                        top : linkClassName,
+                        }}>
+                        {navigationCategories}
                     </div>
-                </header>
-        )
-    }
+                </div>
+            </header>
+    )
 }
